src/requirements.txt
azure-functions
beautifulsoup4
qrcode
Pillow
pytest
azure-storage-blob
